rename: 패키지 구조 레이어드 아키텍처 스타일로 마이그레이션
