+++ src/main/java/com/example/daobe/lounge/application/dto/LoungeCreateRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.daobe.lounge.dto;
+package com.example.daobe.lounge.application.dto;
 
 public record LoungeCreateRequestDto(
         String name,
