+++ src/main/java/com/example/daobe/lounge/application/dto/LoungeCreateResponseDto.java
@@ -1,6 +1,6 @@
-package com.example.daobe.lounge.dto;
+package com.example.daobe.lounge.application.dto;
 
-import com.example.daobe.lounge.entity.Lounge;
+import com.example.daobe.lounge.domain.Lounge;
 
 public record LoungeCreateResponseDto(Long loungeId) {
     public static LoungeCreateResponseDto of(Lounge lounge) {
