+++ src/main/java/com/example/daobe/lounge/application/dto/LoungeInviteDto.java
@@ -1,4 +1,4 @@
-package com.example.daobe.lounge.dto;
+package com.example.daobe.lounge.application.dto;
 
 public record LoungeInviteDto(Long userId, Long loungeId) {
 }
